refactoring moving list-view and UserDirs to a seperate Components
